Implementation/BackEnd/requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
drf-nested-routers
django-allauth
django-rest-auth